refactored the server components and routes
