+++ teamserver/internal/server/server.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	"fmt"
-	"net/http"
 
 	"github.com/gin-gonic/gin"
 	_ "github.com/joho/godotenv/autoload"
@@ -10,6 +9,7 @@ import (
 	"github.com/ksel172/Meduza/teamserver/internal/api/handlers"
 	"github.com/ksel172/Meduza/teamserver/internal/models"
 	"github.com/ksel172/Meduza/teamserver/internal/storage/repos"
+	"github.com/mattn/go-colorable"
 )
 
 type DependencyContainer struct {
@@ -32,6 +32,9 @@ type Server struct {
 
 func NewServer(dependencies *DependencyContainer) *Server {
 
+	// it force the console output to be colored.
+	gin.ForceConsoleColor()
+
 	// Declare Server config
 	server := &Server{
 		host:         conf.GetMeduzaServerHostname(),
@@ -40,69 +43,11 @@ func NewServer(dependencies *DependencyContainer) *Server {
 		dependencies: dependencies,
 	}
 
+	gin.DefaultWriter = colorable.NewColorableStdout()
+
 	server.RegisterRoutes()
 
 	return server
-}
-
-func (s *Server) RegisterRoutes() {
-
-	apiGroup := s.engine.Group("/api")
-	{
-		v1Group := apiGroup.Group("/v1")
-		{
-			// Authentication Routes
-			authRoutes := v1Group.Group("/auth")
-			{
-				authRoutes.Use(s.HandleCors())
-				authRoutes.POST("/register", s.dependencies.UserController.AddUsers)
-				authRoutes.POST("/add-admin", s.dependencies.AdminController.CreateAdmin)
-				authRoutes.POST("/login", s.dependencies.AuthController.LoginController)
-				authRoutes.GET("/refresh-token", s.dependencies.AuthController.RefreshTokenController)
-				authRoutes.POST("/logout", s.dependencies.AuthController.LogoutController)
-			}
-
-			adminProtectedRoutes := v1Group.Group("/teamserver")
-			{
-				adminProtectedRoutes.Use(s.HandleCors())
-				adminProtectedRoutes.Use(s.AdminMiddleware())
-				adminProtectedRoutes.GET("/users", s.dependencies.UserController.GetUsers)
-				adminProtectedRoutes.POST("/users", s.dependencies.UserController.AddUsers)
-			}
-
-			agentsGroup := v1Group.Group("/agents")
-			{
-				agentsGroup.GET(fmt.Sprintf("/:%s", models.ParamAgentID), s.dependencies.AgentController.GetAgent)
-				agentsGroup.PUT(fmt.Sprintf("/:%s", models.ParamAgentID), s.dependencies.AgentController.UpdateAgent)
-				agentsGroup.DELETE(fmt.Sprintf("/:%s", models.ParamAgentID), s.dependencies.AgentController.DeleteAgent)
-
-				// Agent Tasks API
-				agentsGroup.GET("/tasks", s.dependencies.AgentController.GetAgentTasks)
-				agentsGroup.POST("/tasks", s.dependencies.AgentController.CreateAgentTask)
-				agentsGroup.DELETE(fmt.Sprintf(":%s/tasks", models.ParamAgentID), s.dependencies.AgentController.DeleteAgentTasks)
-				agentsGroup.DELETE(fmt.Sprintf(":%s/tasks/:%s", models.ParamAgentID, models.ParamTaskID), s.dependencies.AgentController.DeleteAgentTask)
-			}
-
-			checkinGroup := v1Group.Group("/checkin")
-			{
-				checkinGroup.POST("/", s.dependencies.CheckInController.CreateAgent)
-				checkinGroup.GET("/", s.dependencies.CheckInController.GetTasks)
-			}
-
-			listenersGroup := v1Group.Group("/listeners")
-			{
-				listenersGroup.POST("", s.dependencies.ListenersController.CreateListener)
-				listenersGroup.GET("/:id", s.dependencies.ListenersController.GetListener)
-				listenersGroup.PUT("", s.dependencies.ListenersController.UpdateListener)
-				listenersGroup.DELETE(":id", s.dependencies.ListenersController.DeleteListener)
-			}
-		}
-	}
-
-	// Default HelloWorld Handler
-	s.engine.GET("/", func(context *gin.Context) {
-		context.String(http.StatusOK, "Hello, World!")
-	})
 }
 
 func (s *Server) Run() error {
